+++ Tests/RxCocoaTests/Driver+Test.swift
@@ -178,6 +178,20 @@ extension DriverTest {
         XCTAssertEqual(results, [0, 1, 2])
     }
 
+    func testPublishRelayAsDriver() {
+        let hotObservable: PublishRelay<Int> = PublishRelay()
+        let xs = Driver.zip(hotObservable.asDriver(), Driver.of(0, 0, 0)) { x, _ in
+            return x
+        }
+
+        let results = subscribeTwiceOnBackgroundSchedulerAndOnlyOneSubscription(xs, expectationFulfilled: { $0 == 2 }) {
+            hotObservable.accept(1)
+            hotObservable.accept(2)
+        }
+
+        XCTAssertEqual(results, [0, 1, 2])
+    }
+
     func testAsDriver_onErrorJustReturn() {
         let hotObservable = BackgroundThreadPrimitiveHotObservable<Int>()
         let xs = hotObservable.asDriver(onErrorJustReturn: -1)